:art: Reformat and refactor modules fixtures
