+++ src/modules/BlockchainApplication/__fixtures__/index.js
@@ -1,0 +1,2 @@
+export * from './mockApplications';
+export * from './mockDefaultApplication';
