+++ src/modules/BlockchainApplication/store/selectors.test.js
@@ -1,4 +1,4 @@
-import { BLOCKCHAIN_APPLICATIONS_MOCK, MAPPED_BLOCKCHAIN_APPLICATIONS_MOCK } from '../mocks';
+import { mockApplications, mockMappedApplications } from '../__fixtures__';
 import {
   selectPinnedApplications,
   selectApplications,
@@ -7,7 +7,7 @@ import {
 
 describe('Application Explorer selector', () => {
   it('Should return list of pinned applications if action type is triggered', async () => {
-    const pins = BLOCKCHAIN_APPLICATIONS_MOCK.map(({ chainID }) => chainID);
+    const pins = mockApplications.map(({ chainID }) => chainID);
 
     const state = {
       blockchainApplications: { pins },
@@ -17,14 +17,14 @@ describe('Application Explorer selector', () => {
   });
 
   it('Should return list of all applications if action type is triggered', async () => {
-    const state = { blockchainApplications: { applications: MAPPED_BLOCKCHAIN_APPLICATIONS_MOCK } };
+    const state = { blockchainApplications: { applications: mockMappedApplications } };
 
-    expect(selectApplications(state)).toEqual(MAPPED_BLOCKCHAIN_APPLICATIONS_MOCK);
+    expect(selectApplications(state)).toEqual(mockMappedApplications);
   });
 
   it('Should return current application if setCurrentApplication action type is tiggered', async () => {
-    const state = { blockchainApplications: { current: BLOCKCHAIN_APPLICATIONS_MOCK[0] } };
+    const state = { blockchainApplications: { current: mockApplications[0] } };
 
-    expect(selectCurrentApplication(state)).toEqual(BLOCKCHAIN_APPLICATIONS_MOCK[0]);
+    expect(selectCurrentApplication(state)).toEqual(mockApplications[0]);
   });
 });
