+++ src/modules/SendToken/__fixtures__/index.js
@@ -1,0 +1,3 @@
+export * from './mockTokens';
+export * from './mockFeesByPriorities';
+export * from './mockSendTokenResponse';
