+++ src/modules/SendToken/__fixtures__/mockFeesByPriorities.js
@@ -1,0 +1,5 @@
+export const mockFeesByPriorities = {
+  low: 100,
+  medium: 320,
+  high: 500
+};
