+++ src/modules/SendToken/__fixtures__/mockSendTokenResponse.js
@@ -1,0 +1,4 @@
+export const mockSendTokenResponse = {
+  message: 'Transaction payload was successfully passed to the network node',
+  transactionID: 'bfd3521aeddd586f43931b6972b5771e9919e18f2cc91e940a15eacb588ecc6c'
+};
