+++ src/modules/SendToken/mocks/index.js
@@ -1,55 +1,16 @@
-export const TOKENS_MOCK = [
-  {
-    tokenID: '2ba563cf98003d',
-    name: 'LISK',
-    symbol: 'LSK',
-    availableBalance: '10000000',
-    lockedBalances: [
-      {
-        moduleID: '5',
-        amount: '10000'
-      }
-    ]
-  },
-  {
-    tokenID: '2bc561cf98009a',
-    name: 'COLECTI',
-    symbol: 'COL',
-    availableBalance: '4000000',
-    lockedBalances: [
-      {
-        moduleID: '4',
-        amount: '5000'
-      }
-    ]
-  },
-  {
-    tokenID: '3as601cd98117m',
-    name: 'KALIPO',
-    symbol: 'TKN',
-    availableBalance: '2900000',
-    lockedBalances: [
-      {
-        moduleID: '5',
-        amount: '11000'
-      }
-    ]
-  },
-  {
-    tokenID: '1as631cd08007k',
-    name: 'DOCUEDU',
-    symbol: 'DOE',
-    availableBalance: '34000000',
-    lockedBalances: [
-      {
-        moduleID: '4',
-        amount: '12000'
-      }
-    ]
-  },
-];
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { rest } from 'msw';
 
-export const BROADCASTED_TRANSACTION_MOCK = {
-  message: 'Transaction payload was successfully passed to the network node',
-  transactionID: 'bfd3521aeddd586f43931b6972b5771e9919e18f2cc91e940a15eacb588ecc6c'
-};
+import { API_VERSION } from 'utilities/api/constants';
+import { mockSendTokenResponse } from '../__fixtures__';
+
+export const sendTokenMockHandler = rest.post(
+  `http://104.248.241.229:9901/api/${API_VERSION}/transactions`,
+  (_, res, ctx) => {
+    return res(
+      ctx.status(201),
+      ctx.json(mockSendTokenResponse),
+      ctx.delay(20)
+    );
+  }
+);
