+++ src/modules/Transactions/__fixtures__/index.js
@@ -1,0 +1,2 @@
+export * from './mockTransactions';
+export * from './mockGetTransactionsQuery';
