+++ src/modules/Transactions/__fixtures__/mockGetTransactionsQuery.js
@@ -1,0 +1,10 @@
+import { mockTransactions } from './mockTransactions';
+
+export const mockGetTransactionsQuery = {
+  data: mockTransactions,
+  meta: {
+    count: 2,
+    offset: 0,
+    total: mockTransactions.length,
+  },
+};
