+++ src/utilities/api/__fixtures__/index.js
@@ -1,0 +1,1 @@
+export * from './mockCustomInfiniteQuery';
